alterei os campos para LocalDate
