+++ Back-End/src/main/java/br/com/queerfy/backend/entities/Lease.java
@@ -1,8 +1,10 @@
 package br.com.queerfy.backend.entities;
 
 import br.com.queerfy.backend.dto.LeaseDTO;
+import org.apache.tomcat.jni.Local;
 
 import javax.persistence.*;
+import java.time.LocalDate;
 import java.util.Date;
 
 @Entity
@@ -19,9 +21,9 @@ public class Lease
     @JoinColumn(name = "user_id", referencedColumnName = "id")
     private User user;
     @Column(columnDefinition = "TIMESTAMP WITHOUT TIME ZONE")
-    private Date checkIn;
+    private LocalDate checkIn;
     @Column(columnDefinition = "TIMESTAMP WITHOUT TIME ZONE")
-    private Date checkOut;
+    private LocalDate checkOut;
     private Double totalValue;
     private String leaseStatus;
 
@@ -60,19 +62,19 @@ public class Lease
         this.user = user;
     }
 
-    public Date getCheckIn() {
+    public LocalDate getCheckIn() {
         return checkIn;
     }
 
-    public void setCheckIn(Date checkIn) {
+    public void setCheckIn(LocalDate checkIn) {
         this.checkIn = checkIn;
     }
 
-    public Date getCheckOut() {
+    public LocalDate getCheckOut() {
         return checkOut;
     }
 
-    public void setCheckOut(Date checkOut) {
+    public void setCheckOut(LocalDate checkOut) {
         this.checkOut = checkOut;
     }
