+++ Back-End/src/main/resources/data.sql
@@ -2,12 +2,12 @@ INSERT INTO users (name, birth_date, rg, cpf, email, password) values ('Rafa', '
 INSERT INTO users (name, birth_date, rg, cpf, email, password) values ('Igor', '1999-10-15', '99999', '32321321321312', 'igor@email.com', 'senha')
 INSERT INTO users (name, birth_date, rg, cpf, email, password) values ('Lemos', '1999-10-15', '99999', '32321321321312', 'lemos@email.com', 'senha')
 
-insert into property(name, city,description, likes,active, check_in, check_out, daily_price, filter_date, house_img, latitude, longitude, fk_user) values ('linda casa na praia', 'sao paulo' ,'Ao lado da padaria do bolsonaro','10000' ,'true', '2021-10-15', '2021-10-20', '90.00', 'weekend', 'house_img', '-31321332', '-32183123', '1' )
-insert into property(name, city,description, likes,active, check_in, check_out, daily_price, filter_date, house_img, latitude, longitude, fk_user) values ('linda casa em alphaville', 'penha' ,'Ao lado da mansao do conde','10000' ,'true', '2021-10-15', '2021-10-20', '90.00', 'weekend', 'house_img', '-31321332', '-32183123', '2' )
-insert into property(name, city,description, likes,active, check_in, check_out, daily_price, filter_date, house_img, latitude, longitude, fk_user) values ('casa meio capenga na zona sul', 'penha' ,'Local sujeito a assaltos','10000' ,'true', '2021-10-15', '2021-10-20', '90.00', 'weekend', 'house_img', '-31321332', '-32183123', '3' )
-insert into property(name, city,description, likes,active, check_in, check_out, daily_price, filter_date, house_img, latitude, longitude, fk_user) values ('linda casa na praia', 'rio de janeiro' ,'Ao lado da padaria do bolsonaro','1000' ,'true', '2021-10-15', '2021-10-20', '90.00', 'weekend', 'house_img', '-31321332', '-32183123', '1' )
-insert into property(name, city,description, likes,active, check_in, check_out, daily_price, filter_date, house_img, latitude, longitude, fk_user) values ('linda casa em alphaville', 'rio de janeiro' ,'Ao lado da mansao do conde','1000' ,'true', '2021-10-15', '2021-10-20', '90.00', 'weekend', 'house_img', '-31321332', '-32183123', '2' )
-insert into property(name, city,description, likes,active, check_in, check_out, daily_price, filter_date, house_img, latitude, longitude, fk_user) values ('casa meio capenga na zona sul','penha' , 'Local sujeito a assaltos','1000' ,'true', '2021-10-15', '2021-10-20', '90.00', 'weekend', 'house_img', '-31321332', '-32183123', '3' )
-insert into property(name, city,description, likes,active, check_in, check_out, daily_price, filter_date, house_img, latitude, longitude, fk_user) values ('linda casa na praia', 'sao paulo' ,'Ao lado da padaria do bolsonaro','1000' ,'true', '2021-10-15', '2021-10-20', '90.00', 'weekend', 'house_img', '-31321332', '-32183123', '1' )
-insert into property(name, city,description, likes,active, check_in, check_out, daily_price, filter_date, house_img, latitude, longitude, fk_user) values ('linda casa em alphaville', 'sao paulo' ,'Ao lado da mansao do conde','1000' ,'true', '2021-10-15', '2021-10-20', '90.00', 'weekend', 'house_img', '-31321332', '-32183123', '2' )
-insert into property(name, city,description, likes,active, check_in, check_out, daily_price, filter_date, house_img, latitude, longitude, fk_user) values ('casa meio capenga na zona sul', 'sao paulo' ,'Local sujeito a assaltos','1000' ,'true', '2021-10-15', '2021-10-20', '90.00', 'weekend', 'house_img', '-31321332', '-32183123', '3' )
+insert into property(name, city,description, likes,active,  daily_price, filter_date, house_img, latitude, longitude, fk_user) values ('linda casa na praia', 'sao paulo' ,'Ao lado da padaria do bolsonaro','10000' ,'true',  '90.00', 'weekend', 'house_img', '-31321332', '-32183123', '1' )
+insert into property(name, city,description, likes,active,  daily_price, filter_date, house_img, latitude, longitude, fk_user) values ('linda casa em alphaville', 'penha' ,'Ao lado da mansao do conde','10000' ,'true',  '90.00', 'weekend', 'house_img', '-31321332', '-32183123', '2' )
+insert into property(name, city,description, likes,active,  daily_price, filter_date, house_img, latitude, longitude, fk_user) values ('casa meio capenga na zona sul', 'penha' ,'Local sujeito a assaltos','10000' ,'true',  '90.00', 'weekend', 'house_img', '-31321332', '-32183123', '3' )
+insert into property(name, city,description, likes,active,  daily_price, filter_date, house_img, latitude, longitude, fk_user) values ('linda casa na praia', 'rio de janeiro' ,'Ao lado da padaria do bolsonaro','1000' ,'true',  '90.00', 'weekend', 'house_img', '-31321332', '-32183123', '1' )
+insert into property(name, city,description, likes,active,  daily_price, filter_date, house_img, latitude, longitude, fk_user) values ('linda casa em alphaville', 'rio de janeiro' ,'Ao lado da mansao do conde','1000' ,'true',  '90.00', 'weekend', 'house_img', '-31321332', '-32183123', '2' )
+insert into property(name, city,description, likes,active,  daily_price, filter_date, house_img, latitude, longitude, fk_user) values ('casa meio capenga na zona sul','penha' , 'Local sujeito a assaltos','1000' ,'true',  '90.00', 'weekend', 'house_img', '-31321332', '-32183123', '3' )
+insert into property(name, city,description, likes,active,  daily_price, filter_date, house_img, latitude, longitude, fk_user) values ('linda casa na praia', 'sao paulo' ,'Ao lado da padaria do bolsonaro','1000' ,'true',  '90.00', 'weekend', 'house_img', '-31321332', '-32183123', '1' )
+insert into property(name, city,description, likes,active,  daily_price, filter_date, house_img, latitude, longitude, fk_user) values ('linda casa em alphaville', 'sao paulo' ,'Ao lado da mansao do conde','1000' ,'true',  '90.00', 'weekend', 'house_img', '-31321332', '-32183123', '2' )
+insert into property(name, city,description, likes,active,  daily_price, filter_date, house_img, latitude, longitude, fk_user) values ('casa meio capenga na zona sul', 'sao paulo' ,'Local sujeito a assaltos','1000' ,'true',  '90.00', 'weekend', 'house_img', '-31321332', '-32183123', '3' )
